chore: revert gnl change
