--- lib/lgnl/inc/get_next_line.h
@@ -1,26 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   get_next_line.h                                    :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: ltreser <ltreser@student.42.fr>            +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/08/03 23:14:19 by ltreser           #+#    #+#             */
-/*   Updated: 2023/10/29 01:51:07 by ltreser          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#ifndef GET_NEXT_LINE_H
-# define GET_NEXT_LINE_H
-
-# include <unistd.h>
-# include <stdlib.h>
-# ifndef BUFFER_SIZE
-#  define BUFFER_SIZE 10
-# endif
-
-char	*get_next_line(int fd);
-char	*str_cutter(int *start, int *end, char **longstr, int free_me);
-char	*ft_realloc(char **ptr, int size, int i);
-
-#endif
